+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Deprecated Redact client
 - Use `vonage-jwt-jdk:1.0.2` library instead of `nexmo-jwt-jdk:1.0.1`
 - Allow alphanumeric characters for SMS and MMS sender fields in Messages API
-- Made `WhatsappTemplateRequest` parameters more permissive
+- Fixed incorrect restrictions on `WhatsappTemplateRequest` 
+  - Policy is now optional
+  - Default locale is `en`
+  - `parameters` is now `List<String>`
 - Removed dependency on `commons-io` and `commons-lang3`
 - Ensured User-Agent is set in request headers
 - Added Premium text-to-speech flag in `TalkAction` NCCO